rename: ModalDragHandle 및 ModalContent 컴포넌트 경로 수정 - 컨벤션에 맞게 _components
