+++ app/(anon)/@detail/steps/[step-number]/[detail]/_components/ModalDragHandle.tsx
